workflow_lens_client_python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "workflow-lens-client"
version = "0.1.0"
description = "workflow_lens用UDPクライアントライブラリ"
requires-python = ">=3.7"
dependencies = []

[project.optional-dependencies]
dev = ["pytest>=7.0"]
telemetry = ["opentelemetry-api>=1.20"]

[tool.setuptools.packages.find]
where = ["src"]
